Рефакторинг, корректировка БД.
* Добавить в базу Id для стран и валют - номера по ISO. Переделать группировку на основе типов. Добавить в Sector, Country поля Id.

+++ ndp_invest_helper/DataHandlers/Models/FundsAssetsLink.cs
@@ -5,8 +5,8 @@ namespace ndp_invest_helper.Models
 {
     public partial class FundsAssetsLink
     {
-        public long FundSecurityId { get; set; }
-        public long AssetTypeId { get; set; }
+        public int FundSecurityId { get; set; }
+        public int AssetTypeId { get; set; }
         public decimal Part { get; set; }
     }
 }
